+++ src/route.ts
@@ -4,6 +4,7 @@ import * as Comments from "./controllers/comments"
 import * as Documents from "./controllers/documents"
 import * as Files from "./controllers/files"
 import * as Sessions from "./controllers/sessions"
+import * as Timeline from "./controllers/timeline"
 import * as Users from "./controllers/users"
 
 export const router = new Router()
@@ -47,7 +48,7 @@ router.patch("/users", Users.PatchAll)
 router.get("/users/:id/documents", Users.GetDocuments)
 router.get("/users/:id/comments", Users.GetComments)
 
-router.get("/timeline/:page", Documents.GetTimeline)
+router.get("/timeline/:page", Timeline.GetTimeline)
 
 router.get("/documents/:id", Documents.GetOne)
 router.post("/documents", Documents.Post)
